made the tower use the chunkdecorator, organized mixins
