+++ src/main/java/jamdoggie/betterbattletowers/mixins/accessor/TomlAccessor.java
@@ -1,4 +1,4 @@
-package jamdoggie.betterbattletowers.mixins;
+package jamdoggie.betterbattletowers.mixins.accessor;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
