+++ src/main/resources/betterbattletowers.mixins.json
@@ -4,17 +4,17 @@
   "package": "jamdoggie.betterbattletowers.mixins",
   "compatibilityLevel": "JAVA_8",
   "client": [
-    "ChunkProviderStaticMixin",
     "GuidebookPageMobMixin"
   ],
   "server": [
-    "ChunkProviderServerMixin"
   ],
   "injectors": {
     "defaultRequire": 1
   },
   "mixins": [
-    "ChunkDecoratorOverWorldMixinAddTowers",
-    "TomlAccessor"
+    "decorate.ChunkDecoratorHellMixinAddTowers",
+    "decorate.ChunkDecoratorOverWorldMixinAddTowers",
+    "decorate.ChunkDecoratorRetroMixinAddTowers",
+    "accessor.TomlAccessor"
   ]
 }
